Removed stubs from generator and grammar_parser and put the annotations into the corresponding files
